requirements.txt
altair
blinker
Brotli
cachetools
chardet
click
cssselect
deep-translator
duckduckgo_search
feedfinder2
feedparser
fsspec
gitdb
GitPython
googletrans
h11
h2
hpack
hstspreload
httpcore
httpx
huggingface-hub
hyperframe
idna
jieba3k
joblib
lxml
lxml_html_clean
narwhals
newspaper3k
nltk
numpy
pandas
pillow
protobuf
pyarrow
pydeck
python-dotenv
pytz
regex
requests-file
rfc3986
safetensors
sgmllib3k
smmap
socksio
streamlit
tenacity
tinysegmenter
tldextract
tokenizers
toml
tqdm
transformers
tzdata
watchdog
torch
Django>=4.2.0
python-dotenv>=0.19.0
requests>=2.25.1
beautifulsoup4>=4.9.3
newspaper3k>=0.2.8
python-dateutil>=2.8.2
Pillow>=8.3.1
django-bootstrap5>=22.1
aiohttp>=3.8.1
python-multipart>=0.0.5
